backend/requirements.txt
appdirs==1.4.3
asgiref==3.2.7
astroid==2.4.0
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
defusedxml==0.6.0
distlib==0.3.0
dj-database-url==0.5.0
Django>2,<=2.2.1
django-allauth==0.40.0
django-braces==1.13.0
django-cors-headers==3.1.1
django-decouple==2.1
django-heroku==0.3.1
django-rest-auth==0.9.5
django-rest-framework==0.1.0
django-sendgrid-v5==0.7.1
djangorestframework==3.10.3
docutils==0.16
filelock==3.0.12
future==0.18.2
gunicorn==20.0.4
idna==2.9
isort==4.3.21
jmespath==0.9.5
lazy-object-proxy==1.4.3
mccabe==0.6.1
misaka==2.1.1
mock==4.0.2
oauthlib==3.1.0
pathspec==0.8.0
pbr==5.4.5
pep8==1.7.1
Pillow==7.1.2
psycopg2==2.8.5
psycopg2-binary==2.8.5
pycodestyle==2.5.0
pycparser==2.20
pylint==2.5.0
python-dateutil==2.8.1
python-decouple==3.3
python-http-client==3.2.7
python3-openid==3.1.0
pytz==2020.1
PyYAML==5.3.1
requests==2.23.0
requests-oauthlib==1.3.0
semantic-version==2.8.5
sendgrid==6.3.0
six==1.14.0
sqlparse==0.3.1
termcolor==1.1.0
toml==0.10.0
typed-ast==1.4.1
urllib3==1.25.9
virtualenv==20.0.18
wcwidth==0.1.9
whitenoise==5.0.1
wrapt==1.12.1
